requirements-lock.txt
aichat==0.0.5
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==3.0.0
async-lru==2.0.5
async-timeout==5.0.1
attrs==25.3.0
Authlib==1.6.0
babel==2.17.0
beautifulsoup4==4.13.4
bleach==6.2.0
blinker==1.5
blis==1.3.0
build==1.2.2.post1
catalogue==2.0.10
certifi==2025.6.15
cffi==1.17.1
chardet==5.1.0
charset-normalizer==3.4.2
-e git+ssh://git@github.com/triepod-ai/claude-cli-mcp.git@2f4a736e5afe7de852ca56a24f8eaf833a277b0d#egg=claude_cli_mcp
click==8.2.1
cloudpathlib==0.21.1
comm==0.2.2
confection==0.1.5
confusable-homoglyphs==3.3.1
contourpy==1.3.2
coverage==7.9.1
cryptography==38.0.4
cycler==0.12.1
cymem==2.0.11
dbus-python==1.3.2
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
distro==1.8.0
Django==5.2.3
exceptiongroup==1.3.0
executing==2.2.0
exrex==0.12.0
fastjsonschema==2.21.1
fastmcp==2.8.1
fonttools==4.58.4
fqdn==1.5.1
future==1.0.0
fuzzywuzzy==0.18.0
gensim==4.3.2
gyp==0.1
h11==0.16.0
httpcore==1.0.9
httplib2==0.20.4
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
iniconfig==2.1.0
ipykernel==6.29.5
ipython==9.3.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.1
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
lazr.restfulclient==0.14.5
lazr.uri==1.0.6
Levenshtein==0.27.1
Mako==1.2.4.dev0
marisa-trie==1.2.1
Markdown==3.4.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mcp==1.9.4
mdurl==0.1.2
mistune==3.1.3
murmurhash==1.0.13
narwhals==1.43.1
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
nltk==3.9.1
notebook==7.4.3
notebook_shim==0.2.4
numpy==2.3.1
oauthlib==3.2.2
openapi-pydantic==0.5.1
overrides==7.7.0
packaging==25.0
pandas==2.3.0
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.8
plotly==6.1.2
pluggy==1.6.0
preshed==3.0.10
prometheus_client==0.22.1
prompt_toolkit==3.0.51
psutil==7.0.0
ptyprocess==0.7.0
pugnlp==0.2.6
pure_eval==0.2.3
py-cpuinfo==9.0.0
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.0
pydantic_core==2.33.2
Pygments==2.19.2
PyGObject==3.42.2
PyJWT==2.6.0
pypandoc==1.15
pyparsing==3.0.9
pyproject_hooks==1.2.0
pytest==8.4.1
pytest-asyncio==1.0.0
pytest-benchmark==5.1.0
python-apt==2.6.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-json-logger==3.3.0
python-Levenshtein==0.27.1
python-multipart==0.0.20
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.0
RapidFuzz==3.13.0
redis==6.2.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.0.0
rpds-py==0.25.1
scikit-learn==1.7.0
scipy==1.15.3
seaborn==0.13.2
Send2Trash==1.8.3
shellingham==1.5.4
six==1.16.0
smart-open==7.1.0
sniffio==1.3.1
soupsieve==2.7
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.3
srsly==2.5.1
sse-starlette==2.3.6
stack-data==0.6.3
starlette==0.47.1
terminado==0.18.1
text-unidecode==1.3
thinc==8.3.6
threadpoolctl==3.6.0
tinycss2==1.4.0
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
typer==0.16.0
types-python-dateutil==2.9.0.20250516
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
unicodedata2==16.0.0
uri-template==1.3.0
urllib3==2.4.0
uvicorn==0.34.3
wadllib==1.3.6
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
widgetsnbextension==4.0.14
wrapt==1.17.2
